refactor: move cloud-only server modules to server/src/ee/
